requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
Requests==2.31.0
supabase==1.0.3
markdown==3.4.3
jwt===1.3.1
